+++ bravado/client.py
@@ -66,21 +66,25 @@ from bravado.warning import warn_for_deprecated_op
 log = logging.getLogger(__name__)
 
 
+CONFIG_DEFAULTS = {
+    # See the constructor of :class:`bravado.http_future.HttpFuture` for an
+    # in depth explanation of what this means.
+    'also_return_response': False
+}
+
+
 class SwaggerClient(object):
     """A client for accessing a Swagger-documented RESTful service.
-    """
 
+    :param swagger_spec: :class:`bravado_core.spec.Spec`
+    """
     def __init__(self, swagger_spec):
-        """
-        :param swagger_spec: :class:`bravado_core.spec.Spec`
-        """
         self.swagger_spec = swagger_spec
 
     @classmethod
     def from_url(cls, spec_url, http_client=None, request_headers=None,
                  config=None):
-        """
-        Build a :class:`SwaggerClient` from a url to the Swagger
+        """Build a :class:`SwaggerClient` from a url to the Swagger
         specification for a RESTful API.
 
         :param spec_url: url pointing at the swagger API specification
@@ -89,11 +93,10 @@ class SwaggerClient(object):
         :type  http_client: :class:`bravado.http_client.HttpClient`
         :param request_headers: Headers to pass with http requests
         :type  request_headers: dict
-        :param config: bravado_core config dict. See
-            bravado_core.spec.CONFIG_DEFAULTS
+        :param config: Config dict for bravado and bravado_core.
+            See CONFIG_DEFAULTS in :module:`bravado_core.spec`.
+            See CONFIG_DEFAULTS in :module:`bravado.client`.
         """
-        # TODO: better way to customize the request for api calls, so we don't
-        #       have to add new kwargs for everything
         log.debug(u"Loading from %s" % spec_url)
         http_client = http_client or RequestsClient()
         loader = Loader(http_client, request_headers=request_headers)
@@ -104,7 +107,7 @@ class SwaggerClient(object):
     def from_spec(cls, spec_dict, origin_url=None, http_client=None,
                   config=None):
         """
-        Build a :class:`SwaggerClient` from swagger api docs
+        Build a :class:`SwaggerClient` from a Swagger spec in dict form.
 
         :param spec_dict: a dict with a Swagger spec in json-like form
         :param origin_url: the url used to retrieve the spec_dict
@@ -112,6 +115,10 @@ class SwaggerClient(object):
         :param config: Configuration dict - see spec.CONFIG_DEFAULTS
         """
         http_client = http_client or RequestsClient()
+
+        # Apply bravado config defaults
+        config = dict(CONFIG_DEFAULTS, **(config or {}))
+
         swagger_spec = Spec.from_dict(
             spec_dict, origin_url, http_client, config)
         return cls(swagger_spec)
@@ -248,19 +255,23 @@ class CallableOperation(object):
         warn_for_deprecated_op(self.operation)
         request_params = self.construct_request(**op_kwargs)
         callback = functools.partial(response_callback, operation=self)
-        return self.operation.swagger_spec.http_client.request(request_params,
-                                                               callback)
+        also_return_response = \
+            self.operation.swagger_spec.config['also_return_response']
+        return self.operation.swagger_spec.http_client.request(
+            request_params,
+            callback,
+            also_return_response)
 
 
 def response_callback(incoming_response, operation):
     """
     So the http_client is finished with its part of processing the response.
     This hands the response over to bravado_core for validation and
-    unmarshalling.
+    unmarshalling. On success, the swagger_result is available as
+    `incoming_response.swagger_result`.
 
     :type incoming_response: :class:`bravado_core.response.IncomingResponse`
     :type operation: :class:`bravado_core.operation.Operation`
-    :return: Response spec's return value.
     :raises: HTTPError
         - On 5XX status code, the HTTPError has minimal information.
         - On non-2XX status code with no matching response, the HTTPError
@@ -271,15 +282,16 @@ def response_callback(incoming_response, operation):
     raise_on_unexpected(incoming_response)
 
     try:
-        swagger_return_value = unmarshal_response(incoming_response, operation)
+        incoming_response.swagger_result = unmarshal_response(
+            incoming_response,
+            operation)
     except MatchingResponseNotFound as e:
         six.reraise(
             HTTPError,
             HTTPError(response=incoming_response, message=str(e)),
             sys.exc_info()[2])
 
-    raise_on_expected(incoming_response, swagger_return_value)
-    return swagger_return_value
+    raise_on_expected(incoming_response)
 
 
 def raise_on_unexpected(http_response):
@@ -293,17 +305,15 @@ def raise_on_unexpected(http_response):
         raise HTTPError(response=http_response)
 
 
-def raise_on_expected(http_response, swagger_return_value):
+def raise_on_expected(http_response):
     """
     Raise an HTTPError if the response is non-2XX and matches a response in the
     swagger spec.
 
     :param http_response: :class:`bravado_core.response.IncomingResponse`
-    :param swagger_return_value: The return value of a swagger response if it
-        has one, None otherwise.
     :raises: HTTPError
     """
     if not 200 <= http_response.status_code < 300:
         raise HTTPError(
             response=http_response,
-            swagger_result=swagger_return_value)
+            swagger_result=http_response.swagger_result)