+++ bravado/fido_client.py
@@ -48,7 +48,8 @@ class FidoClient(HttpClient):
     """Fido (Asynchronous) HTTP client implementation.
     """
 
-    def request(self, request_params, response_callback=None):
+    def request(self, request_params, response_callback=None,
+                also_return_response=False):
         """Sets up the request params as per Twisted Agent needs.
         Sets up crochet and triggers the API request in background
 
@@ -57,6 +58,8 @@ class FidoClient(HttpClient):
         :param response_callback: Function to be called after
         receiving the response
         :type response_callback: method
+        :param also_return_response: Consult the constructor documentation for
+            :class:`bravado.http_future.HttpFuture`.
 
         :rtype: :class: `bravado_core.http_future.HttpFuture`
         """
@@ -77,7 +80,8 @@ class FidoClient(HttpClient):
 
         return HttpFuture(concurrent_future,
                           FidoResponseAdapter,
-                          response_callback)
+                          response_callback,
+                          also_return_response)
 
 
 def stringify_body(request_params):